refactor: share qt cpp metadata parsing
